Website 2.0
This is a backup of the old vpchc.org website about 98% before I ever
touched it. The 2% is from me changing every instance of the old company
name, "Vermillion-Park Community Health Center", to "Valley Professional
Community Health Center". I didn't have the time to revert the changes
so it stays.